policies repacked to stop spamming yum update, added mehanism to force priority size

+++ java-1.8.0-openjdk.spec
@@ -86,8 +86,8 @@
 %global aarch64_updatever 40
 %global aarch64_buildver b12
 %global aarch64_changesetid aarch64-hs3135441ed942
-# priority must be 6 digits in total
-%global priority        000000
+# priority must be 7 digits in total
+%global priority        0000000
 %global javaver         1.8.0
 
 # Standard JPackage directories and symbolic links.
@@ -128,7 +128,7 @@
 
 Name:    java-%{javaver}-%{origin}
 Version: %{javaver}.%{updatever}
-Release: 1.%{buildver}%{?dist}
+Release: 4.%{buildver}%{?dist}
 # java-1.5.0-ibm from jpackage.org set Epoch to 1 for unknown reasons,
 # and this change was brought into RHEL-4.  java-1.5.0-ibm packages
 # also included the epoch in their virtual provides.  This created a
@@ -172,6 +172,8 @@ Source12: remove-intree-libraries.sh
 
 # Ensure we aren't using the limited crypto policy
 Source13: TestCryptoLevel.java
+
+Source20: repackReproduciblePolycies.sh
 
 # New versions of config files with aarch64 support. This is not upstream yet.
 Source100: config.guess
@@ -403,6 +405,11 @@ need to.
 
 %prep
 %setup -q -c -n %{name} -T -a 0
+prioritylength=`expr length %{priority}`
+if [ $prioritylength -ne 7 ] ; then
+ echo "priority must be 7 digits in total, violated"
+ exit 14
+fi
 %ifarch %{aarch64}
 pushd jdk8
 rm -r hotspot
@@ -799,6 +806,8 @@ find $RPM_BUILD_ROOT%{_jvmdir}/%{sdkdir}/demo \
     echo "" >> accessibility.properties
   popd
 
+bash %{SOURCE20} $RPM_BUILD_ROOT/%{_jvmdir}/%{jredir} %{javaver}
+
 %post
 update-desktop-database %{_datadir}/applications &> /dev/null || :
 /bin/touch --no-create %{_datadir}/icons/hicolor &>/dev/null || :
@@ -1139,6 +1148,11 @@ exit 0
 %{_jvmdir}/%{jredir}/lib/accessibility.properties
 
 %changelog
+* Thu Feb 12 2015 Jiri Vanek <jvanek@redhat.com> - 1:1.8.0.25-4.b12
+- policies repacked to stop spamming yum update
+- added and used source20 repackReproduciblePolycies.sh
+- added mehanism to force priority size
+
 * Mon Jan 12 2015 Severin Gehwolf <sgehwolf@redhat.com> - 1:1.8.0.31-1.b13
 - Update to January CPU patch update.
